+++ CHANGELOG.md
@@ -24,10 +24,12 @@ Changelog
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 2.1.15 BETA**
+
 * Added restrictions for NEW_PICTURE and ACTION_NEW_VIDEO broadcasts
 * Fixed on demand once caching restriction
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.1.14 BETA**